deployment scripts

+++ src/solc_0.7/catalyst/Gem.sol
@@ -4,13 +4,13 @@ pragma solidity 0.7.1;
 import "./ERC20Token.sol";
 
 contract Gem is ERC20Token {
-    uint256 public immutable gemId;
+    uint16 public immutable gemId;
 
     constructor(
         string memory name,
         string memory symbol,
         address admin,
-        uint256 _gemId
+        uint16 _gemId
     ) ERC20Token(name, symbol, admin) {
         gemId = _gemId;
     }
